NER_and_Image_Classification/requirements.txt
torch
torchvision
transformers
pillow
datasets
numpy
seqeval
